feature/VTA-741: fix unit tests and refactor of logs provider
