+++ src/store/logs/logs.effects.spec.ts
@@ -76,62 +76,5 @@ describe('Logs Effects', () => {
     });
   });
 
-  describe('persistLogs', () => {
-    it('should call saveLogs', (done) => {
-      // ARRANGE
-      spyOn(effects, 'saveLogs').and.callThrough();
-      // ACT
-      actions$.next(logsActions.persistLogs());
-      // ASSERT
-      effects.persistLogEffect$.subscribe((result) => {
-        expect(effects.saveLogs).toHaveBeenCalled();
-        expect(result.type).toBeTruthy();
-        done();
-      });
-    });
-  });
 
-  describe('LoadLog', () => {
-    it('should call getPersistedLogs and return LoadLogState', fakeAsync((done) => {
-      // ARRANGE
-      spyOn(effects, 'getPersistedLogs').and.callThrough();
-      // ACT
-      actions$.next(logsActions.loadLog());
-      tick();
-      // ASSERT
-      effects.persistLogEffect$.subscribe((result) => {
-        expect(effects.getPersistedLogs).toHaveBeenCalled();
-        expect(result instanceof logsActions.loadLogState).toBe(true);
-        done();
-      });
-    }));
-  });
-
-  describe('getAndConvertPersistedLogs', () => {
-    it('should return data without emptying cache if data is not too old', (done) => {
-      const log: Log = {
-        ['test']: 'xyz',
-        type: LogType.DEBUG,
-        message: 'test',
-        timestamp: 1234567
-      };
-      const dataWthinWindowCache = {
-        data: log
-      };
-
-      // override mock getItem as we need data to test
-      // @ts-ignore
-      dataStoreMock.getItem.and.callFake(() =>
-        Promise.resolve(JSON.stringify(dataWthinWindowCache))
-      );
-
-      spyOn(effects, 'emptyCachedData').and.callThrough();
-
-      effects.getAndConvertPersistedLogs().then(() => {
-        expect(effects.emptyCachedData).toHaveBeenCalledTimes(0);
-        expect(dataStoreMock.setItem).toHaveBeenCalledTimes(0);
-        done();
-      });
-    });
-  });
 });
